% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alex_na_plot.R
\name{alex_na_plot}
\alias{alex_na_plot}
\title{Prepares a Graph Displaying Missing Values in Panel Data}
\usage{
alex_na_plot(
  df,
  ts_id,
  no_factors = FALSE,
  binary = FALSE,
  colorx = 10,
  alphax = 1
)
}
\arguments{
\item{df}{Data frame containing the data.}

\item{ts_id}{A string containing the name of the variable indicating the time dimension. Needs to be coercible into an ordered factor.}

\item{no_factors}{A logical value indicating whether you want to limit the plot to
logical and numerical variables. Defaults to \code{FALSE}.}

\item{binary}{If set to \code{TRUE}, the plot uses a binary scale only high-lightening
whether values are missing or not. Defaults to \code{FALSE}.}

\item{colorx}{- color scheme in terms of terrain.colors function from grDevices package (will be loaded and/or installed automatically inside this function) 
The basic value = 10 (10 means: red color for many missings; yellow color for moderate missing; green color for no or almost no missings)}

\item{alphax}{- alpha of the color transparency function in terms of terrain.colors function from grDevices package; by default it is 1.}
}
\value{
A \code{ggplot2} plot. ggplot2, data.table and grDevices packages will be loaded, installed and loaded to your system automatically inside this function - 
whatever your system needs. 
If a package is installed - it will be called; if not - it will be installed and called automatically (user selection for the mirror to download the package may be needed).
}
\description{
Displays a heatmap of missing value frequency across the panel; 
Inspired by the ExPanDaR package approach as a basis
}
\details{
alex_na_plot


  This was inspired by a
  \href{http://www.rensenieuwenhuis.nl/r-sessions-30-visualizing-missing-values/}{blog post of Rense Nieuwenhuis}. And by ExPanDaR package :)

  Thanks! You are the best!

This function has a useful add-on to adjust the design of the chart and select colors; by default it is intuitive reg-yellow-green scheme. 
Datasets with no missing values may be replicated yelow by default - nothing to compare with. You may change the color scheme and design any time.
}
\examples{
library(plm); data("Produc", package="plm"); Produc$test = ifelse(Produc$unemp > 5, Produc$unemp, NA) #Adding NAs occasionally to the dataset
alex_na_plot(Produc, ts_id = "year")
}
